add codec testing
