+++ packages/test/src/test-integration-workflows.ts
@@ -8,7 +8,7 @@ import { msToNumber, tsToMs } from '@temporalio/common/lib/time';
 import { TestWorkflowEnvironment } from '@temporalio/testing';
 import { CancelReason } from '@temporalio/worker/lib/activity';
 import * as workflow from '@temporalio/workflow';
-import { defineQuery, defineSignal, setHandler } from '@temporalio/workflow';
+import { defineQuery, defineSignal } from '@temporalio/workflow';
 import { SdkFlags } from '@temporalio/workflow/lib/flags';
 import {
   ActivityCancellationType,
@@ -19,8 +19,6 @@ import {
   TypedSearchAttributes,
   WorkflowExecutionAlreadyStartedError,
 } from '@temporalio/common';
-import { temporal } from '@temporalio/proto';
-import { decodeOptionalSinglePayload } from '@temporalio/common/lib/internal-non-workflow';
 import { signalSchedulingWorkflow } from './activities/helpers';
 import { activityStartedSignal } from './workflows/definitions';
 import * as workflows from './workflows';
@@ -1414,185 +1412,5 @@ test('Workflow can return root workflow', async (t) => {
   await worker.runUntil(async () => {
     const result = await executeWorkflow(rootWorkflow, { workflowId: 'test-root-workflow-length' });
     t.deepEqual(result, 'empty test-root-workflow-length');
-  });
-});
-
-export async function userMetadataWorkflow(): Promise<{
-  currentDetails: string;
-  childWorkflowId: string;
-  childRunId: string;
-}> {
-  let done = false;
-  const signalDef = defineSignal('done');
-  setHandler(
-    signalDef,
-    () => {
-      done = true;
-    },
-    { description: 'signal-desc' }
-  );
-
-  // That workflow should call an activity (with summary)
-  const { activityWithSummary } = workflow.proxyActivities({
-    scheduleToCloseTimeout: '10s',
-    scheduleToStartTimeout: '10s',
-  });
-  await activityWithSummary.executeWithOptions(
-    {
-      summary: 'activity summary',
-      retry: {
-        initialInterval: '1s',
-        maximumAttempts: 5,
-        maximumInterval: '10s',
-      },
-      scheduleToStartTimeout: '5s',
-    },
-    []
-  );
-  const { localActivityWithSummary } = workflow.proxyLocalActivities({ scheduleToCloseTimeout: '10s' });
-  await localActivityWithSummary.executeWithOptions(
-    {
-      summary: 'local activity summary',
-      retry: {
-        maximumAttempts: 2,
-        nonRetryableErrorTypes: ['CustomError'],
-      },
-      scheduleToStartTimeout: '5s',
-    },
-    []
-  );
-  // Timer (with summary)
-  await workflow.sleep(5, { summary: 'timer summary' });
-  // Set current details
-  workflow.setCurrentDetails('current wf details');
-  // Start child workflow
-  const child_handle = await workflow.startChild(completableWorkflow, {
-    args: [false],
-    staticDetails: 'child details',
-    staticSummary: 'child summary',
-  });
-
-  await workflow.condition(() => done);
-  return {
-    currentDetails: workflow.getCurrentDetails(),
-    childWorkflowId: child_handle.workflowId,
-    childRunId: child_handle.firstExecutionRunId,
-  };
-}
-
-test('User metadata on workflow, timer, activity, child', async (t) => {
-  const { createWorker, startWorkflow } = helpers(t);
-  const worker = await createWorker({
-    activities: {
-      async activityWithSummary() {},
-      async localActivityWithSummary() {},
-    },
-  });
-
-  await worker.runUntil(async () => {
-    // Start a workflow with static details
-    const handle = await startWorkflow(userMetadataWorkflow, {
-      staticSummary: 'wf static summary',
-      staticDetails: 'wf static details',
-    });
-    // Describe workflow -> static summary, static details
-    const desc = await handle.describe();
-    t.true((await desc.staticSummary()) === 'wf static summary');
-    t.true((await desc.staticDetails()) === 'wf static details');
-
-    await handle.signal('done');
-    const res = await handle.result();
-    t.true(res.currentDetails === 'current wf details');
-
-    // Get child workflow handle and verify metadata
-    const childHandle = t.context.env.client.workflow.getHandle(res.childWorkflowId, res.childRunId);
-    const childDesc = await childHandle.describe();
-    t.true((await childDesc.staticSummary()) === 'child summary');
-    t.true((await childDesc.staticDetails()) === 'child details');
-
-    // Get history events for main workflow.
-    const resp = await t.context.env.client.workflowService.getWorkflowExecutionHistory({
-      namespace: t.context.env.client.options.namespace,
-      execution: {
-        workflowId: handle.workflowId,
-        runId: handle.firstExecutionRunId,
-      },
-    });
-    for (const event of resp.history?.events ?? []) {
-      if (event.eventType === temporal.api.enums.v1.EventType.EVENT_TYPE_WORKFLOW_EXECUTION_STARTED) {
-        t.deepEqual(
-          await decodeOptionalSinglePayload(
-            t.context.env.client.options.loadedDataConverter,
-            event.userMetadata?.summary
-          ),
-          'wf static summary'
-        );
-        t.deepEqual(
-          await decodeOptionalSinglePayload(
-            t.context.env.client.options.loadedDataConverter,
-            event.userMetadata?.details
-          ),
-          'wf static details'
-        );
-      } else if (event.eventType === temporal.api.enums.v1.EventType.EVENT_TYPE_ACTIVITY_TASK_SCHEDULED) {
-        t.deepEqual(
-          await decodeOptionalSinglePayload(
-            t.context.env.client.options.loadedDataConverter,
-            event.userMetadata?.summary
-          ),
-          'activity summary'
-        );
-        // Assert that the overriden activity options are what we expect.
-        const attrs = event.activityTaskScheduledEventAttributes;
-        t.is(tsToMs(attrs?.scheduleToCloseTimeout), 10000);
-        t.is(tsToMs(attrs?.scheduleToStartTimeout), 5000);
-        const retryPolicy = attrs?.retryPolicy;
-        t.is(retryPolicy?.maximumAttempts, 5);
-        t.is(tsToMs(retryPolicy?.initialInterval), 1000);
-        t.is(tsToMs(retryPolicy?.maximumInterval), 10000);
-      } else if (event.eventType === temporal.api.enums.v1.EventType.EVENT_TYPE_TIMER_STARTED) {
-        t.deepEqual(
-          await decodeOptionalSinglePayload(
-            t.context.env.client.options.loadedDataConverter,
-            event.userMetadata?.summary
-          ),
-          'timer summary'
-        );
-      }
-    }
-    // Get history events for child workflow.
-    const childResp = await t.context.env.client.workflowService.getWorkflowExecutionHistory({
-      namespace: t.context.env.client.options.namespace,
-      execution: {
-        workflowId: res.childWorkflowId,
-        runId: res.childRunId,
-      },
-    });
-
-    for (const event of childResp.history?.events ?? []) {
-      if (event.eventType === temporal.api.enums.v1.EventType.EVENT_TYPE_WORKFLOW_EXECUTION_STARTED) {
-        t.deepEqual(
-          await decodeOptionalSinglePayload(
-            t.context.env.client.options.loadedDataConverter,
-            event.userMetadata?.summary
-          ),
-          'child summary'
-        );
-        t.deepEqual(
-          await decodeOptionalSinglePayload(
-            t.context.env.client.options.loadedDataConverter,
-            event.userMetadata?.details
-          ),
-          'child details'
-        );
-      }
-    }
-    // Run metadata query -> get current details
-    const wfMetadata = (await handle.query('__temporal_workflow_metadata')) as temporal.api.sdk.v1.IWorkflowMetadata;
-    t.deepEqual(wfMetadata.definition?.signalDefinitions?.length, 1);
-    t.deepEqual(wfMetadata.definition?.signalDefinitions?.[0].name, 'done');
-    t.deepEqual(wfMetadata.definition?.signalDefinitions?.[0].description, 'signal-desc');
-    t.deepEqual(wfMetadata.definition?.queryDefinitions?.length, 3); // default queries
-    t.deepEqual(wfMetadata.currentDetails, 'current wf details');
   });
 });
